+++ src/omnisharp/loggingEvents.ts
@@ -11,6 +11,10 @@ import { LaunchTarget } from "./launcher";
 export interface BaseEvent {
 }
 
+export class TelemetryEvent implements BaseEvent {
+    constructor(public eventName: string, public properties?: { [key: string]: string }, public measures?: { [key: string]: number }){ }
+}
+
 export class TelemetryEventWithMeasures implements BaseEvent {
     constructor(public eventName: string, public measures: { [key: string]: number }) {
     }
@@ -138,7 +142,6 @@ export class DotNetTestDebugProcessStart implements BaseEvent {
     constructor(public targetProcessId: number) { }
 }
 
-
 export class DotNetTestsInClassRunStart implements BaseEvent {
     constructor(public className: string) { }
 }
@@ -149,6 +152,10 @@ export class DotNetTestsInClassDebugStart implements BaseEvent {
 
 export class DocumentSynchronizationFailure implements BaseEvent {
     constructor(public documentPath: string, public errorMessage: string) { }
+}
+
+export class OpenURL {
+    constructor(public url: string) { }
 }
 
 export class DebuggerPrerequisiteFailure extends EventWithMessage { }